feat: Add blocking API in Accessor
